requirements.txt
Pillow>=10.0.0
pyphen>=0.14.0